# Conflicts:
#	src/main/java/dao/OrderDetailDAO.java
#	src/main/java/dao/ProductDAO.java
#	src/main/java/model/Medicine.java
#	src/main/java/service/impl/ProductServiceImpl.java

+++ src/main/java/ui/model/ModelDataPS_Circle.java
@@ -9,9 +9,9 @@ import java.io.Serializable;
 @ToString
 public class ModelDataPS_Circle implements Serializable {
     String name;
-    int count;
+    long count;
 
-    public ModelDataPS_Circle(String name, int count) {
+    public ModelDataPS_Circle(String name, long count) {
         this.name = name;
         this.count = count;
     }
